Add files via upload
Version 1.1.0 released

This version is significant update of RBseq_Count_BarCodes.py.  The readme file and example metafile for this script have been updated to reflect the change.  

The new metafile requires and extra field named FileIndex with a unique index for every fastq file passed through the script.  This new field is required because the script now saves intermediate counts files for each fastq passed, as opposed to saving files aggregating counts over each condition given in the SampleName field.  

In previous versions, RBseq_Count_BarCodes.py aggregated information on ever complaint barcode seen in every fastq file file over the entire run, regardless if that barcode is found in the mapped mutant pool file.  This resulted in the accumulation of very large numbers of low count probable sequence artifact barcodes.  The revised script computes and logs some summary statistics on each fastq processed, and saves counts for all barcodes in the intermediate counts file, but only counts for mapped barcodes are maintained in memory  and aggregated into the poolCounts.txt output.  This change reduces memory usage by an order of magnitude or more on large runs.

When computing summary statistics on the complete distribution of barcodes in each file (mapped or not in the poolfile), this no longer attempts to derive a 'true' distribution of barcode abundances by removing likely sequence errors.  Those calculations were based on too many assumptions and not found to be particularly useful. The script still reports an estimate of barcode population size with the Chao formula.  When sequence depth is limiting and most barcodes sequenced are only seen once or twice, the Chao estimate is probably the LEAST BAD estimate of underlying population size, but it's still pretty bad.  Once a comparable or larger number of barcodes have three or more counts, than one or two, that number is likely a much better indicator of the population of barcodes than the Chao estimate.

The new version also saves a new file fastqSummaryStat.txt that reports all the same summary statistics given on the command line but in a tabular format that is easier to work with for runs with many samples.  These statistics will all still be saved when the -Q/--quietMode flag is passed, but stats requiring information about reads without compliant barcodes won't be saved if previously tabulated counts are loaded for a fastq with the UsePreCounted flag.