Version 1.1.7
A few minor updates.

RBseq_Map_Insertions.py
Barcodes with a high ratio of reads within the insert to reads mapping to genome are no longer removed from the output pool.  These barcodes are more likely to be problematic for a variety of reasons and in older versions if the ratio of reads mapping to the insert was more than seven times the reads mapping to the main location on the genome, the barcode was removed from the pool.  In the current version those barcodes are retained along with the number reads mapping to the genome and the insertion sequence.  This way barcodes with a low number of reads used for mapping or a high ratio of insert reads can be filtered out in the fitness calculation step by the user and thus it's easier to adjust these parameters for optimal results with different mutant pools.

RBseq_Count_BarCodes.py
Minor changes to pass nMainLocation and nInsert to the poolcount file.

RBSeq_Calculate_Fitness.py
(1) New options -N/--minReadsMainLocation and -R/--maximumConcatamerRatio added to allow more control of how some problematic barcodes are excluded from fitness score calculations.
(2) Fixed a bug in which normalization of fitness scores in the first and second half of genes could be thrown off by low insertion numbers in either half producing improperly normalized cor12 plots.
(3) Fixed a minor bug that could prevent proper calculation of insertion density across the largest genome given certain genome naming schemes.